feat(mobile): Design the the main homepage
